requirements/dev.txt
-r base.txt
django-debug-toolbar>=4.2
django-browser-reload>=1.12
django-extensions>=3.2
